+++ app/build.gradle
@@ -39,8 +39,6 @@ def testVersions = [
 
 dependencies {
     compile "com.android.support:appcompat-v7:$versions.support"
-    compile "com.android.support:recyclerview-v7:$versions.support"
-    compile "com.android.support:cardview-v7:$versions.support"
 
     compile "com.squareup.retrofit2:retrofit:$versions.retrofit"
     compile "com.squareup.retrofit2:converter-gson:$versions.retrofit"
