+++ app/src/main/java/com/intive/toz/Pet.java
@@ -1,0 +1,15 @@
+package com.intive.toz;
+/**
+ * Pet model class.
+ */
+public class Pet {
+
+    private String name;
+    private String type;
+    private String gender;
+    private String email;
+    private String address;
+    private String date;
+    private String profilePic;
+
+}
