--- app/src/main/java/com/intive/toz/Retrofit/Pet.java
@@ -1,106 +1,0 @@
-package com.intive.toz.Retrofit;
-/**
- * Pet model class.
- */
-public class Pet {
-
-    private String name;
-    private String type;
-    private String gender;
-    private String email;
-    private String address;
-    private String date;
-    private String profilePic;
-/*
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Pet withName(String name) {
-        this.name = name;
-        return this;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public Pet withType(String type) {
-        this.type = type;
-        return this;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-
-    public Pet withGender(String gender) {
-        this.gender = gender;
-        return this;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public Pet withEmail(String email) {
-        this.email = email;
-        return this;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public Pet withAddress(String address) {
-        this.address = address;
-        return this;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public void setDate(String date) {
-        this.date = date;
-    }
-
-    public Pet withDate(String date) {
-        this.date = date;
-        return this;
-    }
-
-    public String getProfilePic() {
-        return profilePic;
-    }
-
-    public void setProfilePic(String profilePic) {
-        this.profilePic = profilePic;
-    }
-
-    public Pet withProfilePic(String profilePic) {
-        this.profilePic = profilePic;
-        return this;
-    }*/
-
-}
