+++ app/src/main/java/com/intive/toz/network/ApiClient.java
@@ -1,4 +1,4 @@
-package com.intive.toz.Retrofit;
+package com.intive.toz.network;
 
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
@@ -6,18 +6,18 @@ import retrofit2.converter.gson.GsonConverterFactory;
 /**
  * Base url connection.
  */
-public final class RetroClient {
+public final class ApiClient {
 
-    static final String PETS_URL = "http://private-dbfe1-zwierzakiandroid.apiary-mock.com";
+    static final String API_URL = "http://private-dbfe1-zwierzakiandroid.apiary-mock.com";
 
-    private RetroClient() {
+    private ApiClient() {
 
     }
 
     private static Retrofit getRetrofitInstance() {
 
         return new Retrofit.Builder()
-                .baseUrl(PETS_URL)
+                .baseUrl(API_URL)
                 .addConverterFactory(GsonConverterFactory.create())
                 .build();
     }
