+++ app/src/main/res/layout/activity_main.xml
@@ -10,16 +10,5 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="com.intive.toz.MainActivity">
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Hello World!" />
 
-    <Button
-        android:id="@+id/button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"
-        android:text="Button" />
 </RelativeLayout>
